Use dedicated option to download iso and hdd with archive subcommand
Based on the history this did not meant to download other assets than 'other'
assets and it is implemeted using the `asset-size-limit` parameter, which was
set in a quite small size.
https://github.com/os-autoinst/openQA/pull/1571#discussion_r165847808

With this size in place, the error is confusing and not intuitive. It makes
more sense to remove the limit in case you want to get the full assets.

- Introduce `--no-limit` to skip _asset-size-limit_
- by default download only testresults with limits.
- download assets only with an explicit parameter which overrides the size
limits. Side-effects will reset size limits for testresults as well.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>