Allow using a different subdomain via `file_security_policy`
So far the code automatically uses a `.` to separate the specified prefix
which will e.g. redirect from `openqa` to `files.openqa`. With this change
the `.` is no longer implied so one can also specify e.g.
`subdomain:files-` to redirect from e.g. `openqa` to `files-openqa`. The
latter is generally more secure, e.g. when it comes to cookies.

This PR also updates tests and the documentation in the application and
NGINX config to suggest the more secure version.

We might want to introduce `domain:…` in the future to be able to freely
specify a completely different domain but this would require more code
changes.

Related ticket: https://progress.opensuse.org/issues/184456