Automatically increase important durations to match regular durations
This makes it easier to configure storage durations; instead of having to
fix a probable configuration mistake manually it happens automatically with
this change. The disadvantage is that one can now no longer configure a
lower retention for important jobs than for regular jobs but this is most
likely not what anybody wants anyway.

Related ticket: https://progress.opensuse.org/issues/187698