Improve showing advanced fields in case there are warnings
Now that we show warnings (and not just errors) for advanced fields the
code for automatically expanding them needs to be updated to cover
warnings.

The newly introduced code is longer but now we no longer look for a
hard-coded keyword in a human-readable error message so I consider it
cleaner.

Related ticket: https://progress.opensuse.org/issues/187698