Add requirement dependency via symlinks to webui service
Use a symlinks to set dependency to the openqa_webui as it was done in
https://github.com/os-autoinst/openQA/pull/4976/files
in order to restart the webui when the database is restarted.

It seems that something changed since
https://github.com/os-autoinst/openQA/pull/4976 which makes the webui to not
behave as it was at the time of that commit and adding it now ensures that the
sequence of restarts will take place as expected.

poo: https://progress.opensuse.org/issues/183302
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>