Introduce array job settings
With this settings where the key ends with `[]` are split on `,`. So e.g.
`ISSUES[]=foo,bar,baz` will be stored as `ISSUES[]=foo`, `ISSUES[]=bar` and
`ISSUES[]=baz` in the database. So these settings now behave exactly like
`WORKER_CLASS`.

This allows searching for jobs with a certain job setting value more
efficiently. This can now be done via an API query like
`/api/v1/job_settings/jobs?key=ISSUES[]&value=foo`. This query would also
succeed with the values `bar` and `baz` for a job created with
`ISSUES[]=foo,bar,baz`.

Note that such an array setting will still be returned in its joined form
when doing other queries like `/api/v1/jobs/$id` and on the web UI.

Note that the suffix `[]` is still part of the setting key and always needs
to be specified when referring to the setting. That also means that e.g.
`ISSUES` and `ISSUES[]` are two completely different settings.

This change also adjusts the parameter handling of `openqa-clone-job` so
array parameters can be passed there as well.

Related ticket: https://progress.opensuse.org/issues/188652