Emit restart events when job restarted automatically
Send an event through the task itself after the `auto_duplicate` when the job
have been process and send for restart. While the `auto_duplicate` could be
used as a single point for the event's emission it interferes with other
functions which also emit events independently. For instance from the API
restart, it processes the payload and adds data to it, so it left as is.
Using the `restart_openqa_job` from the Restart module should not duplicate
the event which is emitted by the API, which also invoke `auto_duplicate`. And
it should also trigger the event when the job is processed and not while is
queued.

That should make sure that the event is send to AMQP for jobs with
`RETRY` once it is actually triggered.

issue: https://progress.opensuse.org/issues/190557
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>