Emit restart events when job restarted automatically
That should make sure that the event is send to AMQP for jobs with
`RETRY` once it is actually triggered.

Using the `restart_openqa_job` from the Restart module in order to not
duplicate the event which is emitted by the API, which also invoke
`auto_duplicate`. It should also trigger the event when the job is processed
and not while is queued.
To do so, it loads the AMQP plugin to the Gru job. The `one_tick` is
called to trigger the `next_tick` in the `register` function of the AMPQ module
immediately, otherwise it seems like it is registered late and the event is not
emitted.

issue: https://progress.opensuse.org/issues/190557
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>