Retry image optimizations
The optimization can fail with `optipng` returning with a non-zero return
code, see https://progress.opensuse.org/issues/190920. We don't know
exactly why it fails sometimes. Perhaps it is because the image is actually
still written when we invoke `optipng`. This change introduces a simple
retry which should help if it is just that. If `optipng` cannot be executed
at all (e.g. because it isn't installed) the worker will still fail
immediately.

Related ticket: https://progress.opensuse.org/issues/190920