Properly conclude scheduling if there are no jobs
If a wakeup happens while there are no jobs to be scheduled, it did not
emit concluded and the timer for scheduling is left at 0ms, which causes
high CPU and DB use by the scheduler.

Just remove all early exits, which just increased the complexity of the
method without much benefit.