docs: Feature auto-generated deepwiki less prominently
The AI generated deepwiki can be a good entry point for a project
overview but without a guarantee for correctness. We should not feature
it that prominently and as the deepwiki does not have a dynamic status
we should not use a top-level badge. Just a simple link is fine.