Make the number of builds per group on the front page configurable
The number of builds to show per job group on the front page of
the web UI is hardcoded to 3.

SUSE's openQA has a ton of job groups so only showing three per
group makes sense, but not every instance has so many groups, and
the front page can look unnecessarily empty. This allows the
instance maintainer to set an appropriate value.

The value can be overridden for a single query with the
limit_builds query param, but that's unwieldy and doesn't solve
things for casual views of the home page.

Signed-off-by: Adam Williamson <awilliam@redhat.com>