Fix possibly excessive memory use when computer test result overview
When accessing the test results overview page (especially without
parameters, e.g. http://localhost:9526/tests/overview) a very huge set of
jobs is considered. This leads to a very excessive memory usage, e.g. with
current OSD data the RSS rises over 7 GiB and the request takes very long.
There's already a limit of 2000 jobs which is also effective but
unfortunately not passed to `complex_query` and only used later on. I
tracked the memory usage while running the code with debug printing it it
was very clear that `complex_query` is the culprit.

This change uses the limit from the start. This means additional
filtering done in Perl code also needs to happen as part of the now
limited initial query (using a sub query, see the added comment). So this
change replaces these Perl loops/grep with passing the filter parameters
directly to the database. This slightly changes the order of things but
should lead to the same end result.

With this change the "preferred" machine (basically the most frequent
machine per architecture) is only determined based on the jobs that are
actually displayed. That means jobs will no longer show as e.g. `kde@uefi`
when filtering for `machine=uefi` because with this filter the machine
`uefi` becomes the preferred (most frequent) machine (as now all displayed
jobs have that machine). I think this change is acceptable. It means that
some tests had to be adjusted accordingly.

Related ticket: https://progress.opensuse.org/issues/192448