feat: Allow to configure key+secret with env variables
This is a better variant to
https://github.com/os-autoinst/openQA/pull/6767 which only adds the
support in the worker authentication. If we change the user agent then
we can enable clients as well as workers to use environment variables.

This commit adds support for optional environment variables
OPENQA_API_KEY and OPENQA_API_SECRET which if specified take precedence
over config file entries. Command line arguments take precedence over
environment variables but are now explicitly discouraged in help texts
and documentation to reduce the risk of leaking credentials.