feat(OpenQA::Git): Cleanup git dir in commit() on shutdown
Issue: https://progress.opensuse.org/issues/193468

OpenQA::Git::_run_cmd will throw an exception when
* The process received a signal
* or when a git command was aborted due to a signal

OpenQA::Git::commit will attempt to cleanup after itself if aborted by a signal
in the middle.

The delete_needles task will abort and return an error message to the user
if it was stopped by a signal, but the task itself will succeed.